StudyGroup Command service 계층(Impl) 분리
