+++ innerjoinus-studygroup-source/src/main/java/org/thisdote/innerjoinus/studygroupservice/command/controller/StudyGroupCommandController.java
@@ -2,7 +2,6 @@ package org.thisdote.innerjoinus.studygroupservice.command.controller;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
